Update mutable-string-class.java
As string objects are immutable, it is sometimes inconvenient to make some changes to string objects. This MutableString class will act as a container of a string object and allow you to modify the original string.
Also, Java provides the StringBuilder class in case we modify the original string frequently.